# docs/openapi.yaml
openapi: 3.0.3

info:
  title: XA Bank Time Deposit API
  version: 1.0.0
  description: API for retrieving time deposits and updating all time deposit balances.

servers:
  - url: http://localhost:8080
    description: Local development and Docker Compose default
  - url: http://localhost:{port}
    description: Local development with custom server port
    variables:
      port:
        default: '8080'
        description: HTTP port configured by SERVER_PORT

paths:
  /time-deposits:
    get:
      summary: Get time deposits
      operationId: getTimeDeposits
      parameters:
        - name: page
          in: query
          description: Zero-based page index. Defaults to 0.
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Page size. Defaults to 20 and must not exceed 100.
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          description: Sort expression in the form field,direction. Supported fields are id. Defaults to id,asc.
          required: false
          schema:
            type: string
            default: id,asc
            enum:
              - id,asc
              - id,desc
      responses:
        '200':
          description: Page of time deposits
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeDepositPageResponse'

  /time-deposits/balances:
    post:
      summary: Update balances of all time deposits
      operationId: updateTimeDepositBalances
      responses:
        '204':
          description: All time deposit balances were updated successfully

components:
  schemas:
    TimeDepositPageResponse:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TimeDepositResponse'
        page:
          type: integer
          format: int32
          example: 0
        size:
          type: integer
          format: int32
          example: 20
        totalElements:
          type: integer
          format: int64
          example: 125
        totalPages:
          type: integer
          format: int32
          example: 7

    TimeDepositResponse:
      type: object
      required:
        - id
        - planType
        - balance
        - days
        - withdrawals
      properties:
        id:
          type: integer
          format: int32
          example: 1
        planType:
          type: string
          example: premium
        balance:
          type: number
          example: 1050.00
        days:
          type: integer
          format: int32
          example: 60
        withdrawals:
          type: array
          items:
            $ref: '#/components/schemas/WithdrawalResponse'

    WithdrawalResponse:
      type: object
      required:
        - id
        - amount
        - date
      properties:
        id:
          type: integer
          format: int32
          example: 10
        amount:
          type: number
          example: 100.00
        date:
          type: string
          format: date
          example: 2026-09-01